+++ Farmaceutico.java
@@ -3,7 +3,7 @@ public class Farmaceutico extends Medico {
     private String licencia;
     private int preescripcionesPorDia;
 
-    public Farmaceutico(int id, String nombre, String departamento, int experiencia, int salarioBase, String licencia, int preescripcionesPorDia) {
+    public Farmaceutico(int id, String nombre, String departamento, int experiencia, double salarioBase, String licencia, int preescripcionesPorDia) {
         super(id, nombre, departamento, experiencia, salarioBase);
         this.licencia = licencia;
         this.preescripcionesPorDia = preescripcionesPorDia;
